store test results in a mapping rather than as method attributes
